requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psutil==5.9.5
google-api-python-client==2.97.0
google-auth==2.22.0
Werkzeug==2.3.7
python-dotenv==1.0.0